Added phonetic transcriber for remote use cases.
Also includes a small test class. This basically contains two (2) classes, one to be used on the Windows machine acting as a server for the TCL CTT script and one on any other remote client. The clients connect and send the user input to the server which runs the script, and repackages the transcribed phonetic representation and sends it back over the socket to the client.

The server/clients use sockets over INET and currently the server is threaded so it should in theory (not tested) be able to handle multiple requests. Also it should be able to shutdown gracefully, closing sockets after itself. All subthreads (the thread listening to incoming client connections and all further spawned threads handling communication with the clients) are daemon threads and will exit/get killed when the main thread is killed. The main thread intercepts KeyboardInterrupts and it is currently the only way of controlling a shutdown. Look in the code for the close_listener()-function where you can add more clean up/socket closing code.

Signed-off-by: parhamfh <parhamfh@kth.se>